+++ arenatacticsserver/scenes/world/hex_grid_cube.gd
@@ -9,14 +9,21 @@ enum GenerationAlgorithm {GRID, CIRCLE}
 @export var generation_algorithm: GenerationAlgorithm
 ## Delay each tile spawn by this amount (in seconds).
 @export_range(0, 5, 0.01, "or_greater") var tile_spawn_delay: float = 0.0
-## Delay each tile animation by this amount (in seconds).
-@export_range(0, 5, 0.01, "or_greater") var tile_animation_delay: float = 0.025
+
 ## Enable or disable elevation variation from Noise-Based Textures (See [ProceduralGenerator]).
 @export var flat_world: bool = false
 ## Enable or disable grid interactivity with signals.
 @export var interactive_grid: bool = true
 ## The tile that is going to be generated
 @export var tile_scene: PackedScene = preload("res://scenes/world/tiles/HexTileCube.tscn")
+
+@export_subgroup("Animations")
+## Delay each fade animation by this amount (in seconds).
+@export_range(0, 5, 0.01, "or_greater") var fade_animation_delay: float = 0.03
+## Delay each tile animation by this amount (in seconds).
+@export_range(0, 5, 0.01, "or_greater") var tile_animation_delay: float = 0.03
+## Delay each grid circle animation by this amount (in seconds).
+@export_range(0, 5, 0.01, "or_greater") var circle_animation_delay: float = 0.3
 
 @export_subgroup("Grid Generation Settings")
 ## Define the amount of tile per "line", AKA Columns.
@@ -50,6 +57,8 @@ var maximum_x_size: int:
 var maximum_y_size: int:
 	get():
 		return get_maximum_size().y
+
+var circle_rings: Array[Array] = []
 
 signal tile_hovered(hex_tile: HexTileCube)
 signal tile_clicked(hex_tile: HexTileCube)
@@ -139,17 +148,31 @@ func generate_circle_grid() -> void:
 		generate_tile(coord)
 		if tile_spawn_delay > 0:
 			await get_tree().create_timer(tile_spawn_delay).timeout
+	circle_rings = get_rings(center_tile, circle_coordinates)
 	is_generation_finished = true
 	generation_finished.emit()
+	
+
+func get_rings(center_tile: HexTileCube, coordinates: Array[Vector3i]) -> Array[Array]:
+	var rings: Array[Array] = [[center_tile]]
+	var remaining_rings = circle_diameter - 1
+	for i in range(remaining_rings):
+		var tiles_in_ring: Array[HexTileCube] = []
+		for j in range(get_tile_count_per_ring(i + 1)):
+			tiles_in_ring.append(get_tile_at_cube_coords(coordinates.pop_front()))
+		rings.append(tiles_in_ring)
+	return rings
+
+func get_tile_count_per_ring(ring_number: int) -> int:
+	return ring_number * 6
 
 func get_circle_starting_position() -> Vector3i:
 	return HexTileCube.get_cube_coordinate(Vector2i(circle_diameter, circle_diameter))
 
 #enregion
 
-#region GridAnimation
+#region Animation
 func fade(out: bool = true, fast_forward_to_end: bool = false) -> void:
-	print("fading")
 	for tile: HexTileCube in get_children():
 		if tile.anim_player:
 			if out:
@@ -160,28 +183,46 @@ func fade(out: bool = true, fast_forward_to_end: bool = false) -> void:
 				tile.anim_player.play("tile_animation_library/fade_in")
 			if fast_forward_to_end:
 				tile.anim_player.advance(1)
-			if tile_animation_delay > 0 && not fast_forward_to_end:
-				await get_tree().create_timer(tile_animation_delay).timeout
+			if fade_animation_delay > 0 && not fast_forward_to_end:
+				await get_tree().create_timer(fade_animation_delay).timeout
 	await get_tree().create_timer(1.0).timeout
 	if out:
-		print("fade out finished")
 		fade_out_finished.emit()
 	else:
-		print("fade in finished")
 		fade_in_finished.emit()
 
-func queue_grid_anim(anim_name: String, wait_mult: float = 1) -> void:
+## Add an animation to each tile in the hex grid. The animations are played in the generation order and can be delayed with [member HexGridCube.tile_animation_delay] (in seconds).
+func queue_grid_anim(anim_name: String) -> void:
 	for tile: HexTileCube in get_children():
 		if tile.anim_player:
 			if tile.anim_player.has_animation(anim_name):
 				tile.anim_player.queue(anim_name)
 			if tile_animation_delay > 0:
-				await get_tree().create_timer(tile_animation_delay * wait_mult).timeout
+				await get_tree().create_timer(tile_animation_delay).timeout
+
+## Add an animation to each tile in the hex grid.[br]
+## The animations of every tiles in a ring are synchronised and each rings trigger can be delayed with [member HexGridCube.tile_animation_delay] (in seconds).[br]
+## `reverse_order` make the animation start at the outermost ring if `true`.[br]
+## `delay_inside_rings` delay each animation by [member HexGridCube.tile_animation_delay] inside a ring if true.
+func queue_rings_anim(anim_name: String, reverse_order: bool = false, delay_inside_rings: bool = false):
+	var rings: Array[Array] = circle_rings.duplicate()
+	if reverse_order:
+		rings.reverse()
+	for ring: Array[HexTileCube] in rings:
+		for tile: HexTileCube in ring:
+			if tile.anim_player:
+				if tile.anim_player.has_animation(anim_name):
+					tile.anim_player.queue(anim_name)
+				if delay_inside_rings && tile_animation_delay > 0:
+					await get_tree().create_timer(tile_animation_delay).timeout
+		if circle_animation_delay > 0:
+			await get_tree().create_timer(circle_animation_delay).timeout
 
 func update_tiles_animations_colors(colors: Dictionary):
 	var tile: HexTileCube = self.get_children().front()
 	if tile:
 		tile.update_animations_colors(colors)
+
 
 #endregion
 
@@ -203,14 +244,20 @@ func get_center_tile() -> HexTileCube:
 	for tile: HexTileCube in self.get_children():
 		match generation_algorithm:
 			GenerationAlgorithm.CIRCLE:
-				return get_tile_at_coords(Vector2i(circle_diameter, circle_diameter))
+				return get_tile_at_offset_coords(Vector2i(circle_diameter, circle_diameter))
 			_:
-				return get_tile_at_coords(Vector2i(ceil(float(maximum_x_size) / 2), ceil(float(maximum_y_size) / 2)))
+				return get_tile_at_offset_coords(Vector2i(ceil(float(maximum_x_size) / 2), ceil(float(maximum_y_size) / 2)))
 	return null
 
-func get_tile_at_coords(_offset_coordinate: Vector2i) -> HexTileCube:
+func get_tile_at_offset_coords(_offset_coordinate: Vector2i) -> HexTileCube:
 	for tile: HexTileCube in self.get_children():
 		if tile.offset_grid_coordinate == _offset_coordinate:
+			return tile
+	return null
+
+func get_tile_at_cube_coords(_cube_coordinate: Vector3i) -> HexTileCube:
+	for tile: HexTileCube in self.get_children():
+		if tile.grid_coordinate == _cube_coordinate:
 			return tile
 	return null